requirements.txt
torch 
transformers 
datasets 
sentence-transformers 
faiss-gpu 
evaluate 
tqdm 
bitsandbytes 
activeft
python-dotenv